+++ template-parts/classic-blocks/header_sections.php
@@ -1,14 +1,12 @@
 <?php
 /*
  * Different layouts for the header area can be used
- * Add your layouts styles to grid.scss
- * Some image classes can be found in media.scss
  *
  */
 
 
 $id = get_the_ID();
-// check if the flexible content field has rows of data without acf functions
+
 if ( function_exists('have_rows') && have_rows( 'header_layout', $id ) && get_field( 'override_header' ) ):
 	while ( have_rows( 'header_layout', $id ) ): the_row();
 		$header = get_row_layout();
@@ -18,11 +16,12 @@ if ( function_exists('have_rows') && have_rows( 'header_layout', $id ) && get_fi
 	endwhile;
 else:
 
-	// default header as a background.
-	// you can add a bg color on top by adding a gradient into background image in the style below
+
+//DEFAULT HEADER IF NO CUSTOM HEADER CHOSEN
+
 	$bg_image = ign_get_header_image( get_the_ID() );
 	?>
-    <header class="entry-header layout-center-content"
+    <header class="entry-header layout-center-content overlay"
 	        <?php if ( $bg_image ) { ?>style="background-image: url('<?php echo $bg_image; ?>');" <?php } ?>>
 
         <div class="header-content container-fluid text-center">